backend/services/account-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
python-multipart
